Create model classes
In terms of model I mean busieness logic. I realize exchange service in the
Exchange class.

Exchange class operates with linker-list. It puts all current sell orders
in front of this list and all buy orders in the end of this list. If orders intersect
server matches them.

All brokers have their tokens, which store in the exchange class.