Rewrite exchange class logic
It became difficult to write the algorithm to add and sell order via just one linked list.
That's why i decided to make another, more simple, structure:

Now Echange class operates with two sorted lists: one for sell orders and one for
buy orders. These lists are sorted with the following rule: sell increases, buy decreases.

The algorithm is now enough simple. If we make the sell order, firstly it checks the first node
of the buy list. If buy bet is lower we correctly put sell order in the sell list, otherwise
we match orders.