Create handler interface
I created Scneraio class. This class needs to be the one accsess point
to different scenarious.

When handler gets HTTP request it delegates it's handling to the scenario class.
All possible actions are stored in the scenario map.

Moreover, I made some fixes in the model to make request handling easier.